Made %hours% display as an integer instead of a double

+++ src/main/java/me/dave/activityrewarder/module/playtimedailygoals/PlaytimeDailyGoalsModule.java
@@ -142,7 +142,7 @@ public class PlaytimeDailyGoalsModule extends Module {
 
         ChatColorHandler.sendMessage(player, ActivityRewarder.getConfigManager().getMessage("daily-playtime-reward-given")
             .replaceAll("%minutes%", String.valueOf(ActivityRewarder.getDataManager().getRewardUser(player).getMinutesPlayed()))
-            .replaceAll("%hours%", String.valueOf(Math.floor(ActivityRewarder.getDataManager().getRewardUser(player).getMinutesPlayed() / 60D))));
+            .replaceAll("%hours%", String.valueOf((int) Math.floor(ActivityRewarder.getDataManager().getRewardUser(player).getMinutesPlayed() / 60D))));
 
         playtimeDailyGoalsModuleUserData.setLastCollectedPlaytime(totalMinutesPlayed);
         ActivityRewarder.getDataManager().saveRewardUser(rewardUser);
